third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-tap-reset-selection-range.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<div id="dragme">Make Selection Here</div>
<script>
description("This tests if the selection is reset after tapping of gesture");

if (window.eventSender) {
    var dragme = document.getElementById('dragme');
    window.getSelection().selectAllChildren(dragme);
    shouldBeGreaterThan('window.getSelection().rangeCount', '0');

    var x = dragme.offsetLeft + dragme.offsetWidth / 2;
    var y = dragme.offsetTop + dragme.offsetHeight / 2;
    eventSender.gestureTapDown(x, y);
    eventSender.gestureShowPress(x, y);
    eventSender.gestureTap(x, y);
    shouldBeZero('window.getSelection().rangeCount');
}
</script>
