third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-tap-reset-selection-range-expected.txt

This tests if the selection is reset after tapping of gesture

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getSelection().rangeCount is > 0
PASS window.getSelection().rangeCount is 0
PASS successfullyParsed is true

TEST COMPLETE
Make Selection Here
